refactor: extract shared FileBackendExecution for SSH support
Move remote CLI execution, SSH command runner, and CSV parsing from
SQLite-specific code to shared DB/ level abstractions. Any file-based
backend now gets SSH support by using RemoteCLIExecution in connect().

Add DuckDB over SSH support using the same pattern.
